+ CPU clustering algorithm corrected + 64 bit creature id + create random number for cells and particles

+++ source/EngineGpuKernels/Object.cuh
@@ -295,7 +295,7 @@ struct Cell
     bool sticky;
     uint32_t age;
     LivingState livingState;
-    uint32_t creatureId;
+    uint64_t creatureId;
     uint32_t mutationId;
     uint8_t ancestorMutationId;  // Only the first 8 bits from ancestor mutation id
     float genomeComplexity;
